User_TakeTesting/deploy4.5_FixF5From4.3/requirements.txt
pandas
gradio
openai
openpyxl

aiohttp
circuitbreaker

gdown

aiohttp
asyncio
circuitbreaker
gradio
logging
openai
openpyxl
pandas


python-dotenv

aiohttp